+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Label text can be displayed by clicking on it
 
 ### Changed
+- Improve color brightness detection
 - Fix task filter on dashboard when not admin
 
 ### Removed